% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationMALDI_Bot.R
\name{annotationMALDI_Bot}
\alias{annotationMALDI_Bot}
\title{Annotate Bottom-up MALDI-MS experiments}
\usage{
annotationMALDI_Bot(
  ProcessedMALDIMS,
  TheoreticalMass,
  mzw = 1,
  write = FALSE,
  plot = TRUE,
  orig = TRUE,
  lint = 3,
  rint = 3,
  FilterInt = TRUE,
  In = 10,
  plotIndi = TRUE,
  dir = getwd(),
  save.name
)
}
\arguments{
\item{ProcessedMALDIMS}{Output obtain from \code{\link{ProcessMALDIMS}}.}

\item{TheoreticalMass}{Output obtained from \code{\link{Protein.to.Peptide}}.}

\item{mzw}{Mass-to-charge window (Da) for the annotation. Default to 1.}

\item{write}{Write an output table with results. Default to  FALSE.}

\item{plot}{Plot the peptides found. Default to TRUE.}

\item{orig}{Plot raw MS spectra. Default to TRUE.}

\item{lint}{Left interval for plotting purposes. Default to 3.}

\item{rint}{Right interval for plotting purposes. Default to 3.}

\item{FilterInt}{Control filtering by relative intensity. Default to TRUE.}

\item{In}{value for filtering by relative intensity. Default to 10.}

\item{plotIndi}{Plot specific peptides matched. Default to TRUE.}

\item{dir}{Directory where to save plots. Default to C:.}

\item{save.name}{Name to save output.}
}
\value{
Peaks annotated in the MALDI-MS spectra.
}
\description{
Annotate Bottom-up MALDI-MS experiments
}
