% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assign_Mass.R
\name{Assign_Mass}
\alias{Assign_Mass}
\title{Annotate ESI-MS spectra}
\usage{
Assign_Mass(
  Scored,
  sequence,
  Metal,
  mono = FALSE,
  NbindingMetal = 7,
  Maxcharge = 6,
  Adduct = NULL,
  mzw = 1,
  write = FALSE,
  dir = getwd()
)
}
\arguments{
\item{Scored}{Scored ESI-MS spectra}

\item{sequence}{Protein sequence.}

\item{Metal}{Metal ion to include in the calculation of the MS library.}

\item{mono}{If TRUE, monoisotopic mass is used.}

\item{NbindingMetal}{Number of expected metal ions binding to the protein. Default to 7.}

\item{Maxcharge}{Maximum charge expected for adduct considered. Default to 6.}

\item{Adduct}{Adduct for library generation.}

\item{mzw}{Mass error window}

\item{write}{If TRUE, reports the results in csv file}

\item{dir}{Directory where to save the results}
}
\value{
Annotate ESI-MS spectra
}
\description{
Annotate ESI-MS spectra
}
