% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assign_MassB.R
\name{Assign_MassB}
\alias{Assign_MassB}
\title{Annotate ESI-MS spectra}
\usage{
Assign_MassB(
  Scored,
  external = "Zn4MT2A_IAM_low_nobaseline_peaks.dat",
  sequence,
  Alkylator = c("IAM"),
  Metal = c("Zn"),
  mono = FALSE,
  NbindingMetal = 7,
  Adduct = NULL,
  mzw = 1,
  write = FALSE,
  dir = getwd()
)
}
\arguments{
\item{Scored}{Scored ESI-MS spectra}

\item{external}{External mass deconvoluton list from UniDec.}

\item{sequence}{Protein sequence.}

\item{Alkylator}{Cysteine alkylator employed}

\item{Metal}{Metal ion to include in the calculation of the MS library.}

\item{mono}{If TRUE, monoisotopic mass is used.}

\item{NbindingMetal}{Number of expected metal ions binding to the protein. Default to 7.}

\item{Adduct}{Adduct for library generation.}

\item{mzw}{Mass error window}

\item{write}{If TRUE, reports the results in csv file}

\item{dir}{Directory where to save the results}
}
\value{
Annotate ESI-MS spectra
}
\description{
Annotate ESI-MS spectra
}
