% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountHits.R
\name{CountHits}
\alias{CountHits}
\title{Count Hit ratio}
\usage{
CountHits(TheoreticalMass, maldiTOF)
}
\arguments{
\item{TheoreticalMass}{Theoretical massess}

\item{maldiTOF}{Output obtained from \code{\link{annotationMALDI_Bot}}.}
}
\value{
Count Hit ratio for PMF MALDI-MS.
}
\description{
Count Hit ratio
}
