% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolved_ESI_High.R
\name{Deconvolved_ESI_High}
\alias{Deconvolved_ESI_High}
\title{Pre-process and deconvolved ESI-MS spectra with Zscore algorithm}
\usage{
Deconvolved_ESI_High(
  exp,
  maxMass,
  SNR.Th,
  max.charge = 9,
  tol = 0.015,
  mz1 = 1000,
  peak_picking = FALSE
)
}
\arguments{
\item{exp}{Experimental MS file.}

\item{maxMass}{Maximum mass deconvoluted}

\item{SNR.Th}{Signal-to-noise ratio. Default to 3.}

\item{max.charge}{Maximum charge state}

\item{tol}{mass error for deconvolution}

\item{mz1}{Select lower m/z limit.}

\item{peak_picking}{If TRUE, peak picking is performed.}
}
\value{
Processed and deconvolved ESI-MS spectra.
}
\description{
Pre-process and deconvolved ESI-MS spectra with Zscore algorithm
}
