% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoy.R
\name{decoy}
\alias{decoy}
\title{Produce a decoy database}
\usage{
decoy(sequence, Npermu = 1, df_full)
}
\arguments{
\item{sequence}{protein sequence}

\item{Npermu}{Number of permutations}

\item{df_full}{target database}
}
\value{
Decoy database
}
\description{
Produce a decoy database
}
