% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution_plot.R
\name{Distribution_plot}
\alias{Distribution_plot}
\title{Distribution plot}
\usage{
Distribution_plot(
  pmfscore_totalb,
  threshold,
  binwidth = 0.2,
  sd,
  name,
  do.plotB = FALSE
)
}
\arguments{
\item{pmfscore_totalb}{Output of the scoring}

\item{threshold}{threshold for scoring}

\item{binwidth}{Binwidth}

\item{sd}{sd}

\item{name}{name for saving}

\item{do.plotB}{If TRUE returns plot}
}
\value{
Distribution plot for scored PMF MALDI-MS.
}
\description{
Distribution plot
}
