% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELDP.R
\name{ELDP}
\alias{ELDP}
\title{Calculates ELDP}
\usage{
ELDP(maldiTOF, sequence)
}
\arguments{
\item{maldiTOF}{Output obtained from \code{\link{annotationMALDI_Bot}}.}

\item{sequence}{Protein sequence.}
}
\value{
Calculates ELDP for PMF MALDI-MS.
}
\description{
Calculates ELDP
}
