% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESI.theoretical.R
\name{ESI.Theoretical}
\alias{ESI.Theoretical}
\title{ESI-MS library construction for metal-binding protein}
\usage{
ESI.Theoretical(
  sequence,
  Metal = c("Zn"),
  NbindingMetal,
  Maxcharge,
  Adduct = c("NH4")
)
}
\arguments{
\item{sequence}{Protein sequence.}

\item{Metal}{Metal ion to include in the calculation of the MS library.}

\item{NbindingMetal}{Number of expected metal ions binding to the protein. Default to 7.}

\item{Maxcharge}{Maximum charge expected for adduct considered. Default to 6.}

\item{Adduct}{Adduct for library generation.}
}
\value{
ESI-MS library for metal-binding protein.
}
\description{
Generates a theoretical ESI-MS library for a particular metal binding protein selecting the desired metal ion, metal ion stoichiometry and adducts.
}
\examples{
\dontrun{
MT2a<-("MDPNCSCAAGDSCTCAGSCKCKECKCTSCKKSCCSCCPVGCAKCAQGCICKGASDKCSCCA")
Theoretical.esi<-ESI.Theoretical(MT2a, Metal=c("Zn"), NbindingMetal = 7,Maxcharge=6,Adduct=c("Na"))
}
}
