% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitPeakToGaussian.R
\name{FitPeakToGaussian}
\alias{FitPeakToGaussian}
\title{FitpeakToGaussian}
\usage{
FitPeakToGaussian(
  x,
  y,
  centroid,
  doPlot = TRUE,
  pTitle = "Peak",
  xTitle = "x",
  yTitle = "y"
)
}
\arguments{
\item{x}{vector with the massess}

\item{y}{vector with the intensity}

\item{centroid}{mean of the peak}

\item{doPlot}{- A boolean (default FALSE) to plot the data, estimate, and the fit.}

\item{pTitle}{- An optional plot title}

\item{xTitle}{- An optional X-axis label}

\item{yTitle}{- An optional Y-axis label}
}
\value{
sum - A summary of the fit
}
\description{
Using the output from EstimateGaussianParameters, do a nonlinear
fit to a Gaussian model
}
\examples{
# Not run
}
