% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldiTOF.Amp.R
\name{maldiTOF.Amp}
\alias{maldiTOF.Amp}
\title{Extended annotation for Bottom-up MALDI-MS experiments}
\usage{
maldiTOF.Amp(sequence, maldiTOF, missed = 2, enzyme = "trypsin", write = TRUE)
}
\arguments{
\item{sequence}{Protein sequence.}

\item{maldiTOF}{Output obtained from \code{\link{annotationMALDI_Bot}}.}

\item{missed}{Missed cleavage sites produced by proteolytic enzyme. Default to 2.}

\item{enzyme}{Proteolytic enzyme used in the experiment. Default to Trypsin.}

\item{write}{Write results. Default to TRUE.}
}
\value{
Extended annotation for bottom-up MALDI-MS experiment.
}
\description{
Extended annotation for Bottom-up MALDI-MS experiments including the start, stop and missed cleavage sites from the proteolytic reaction.
}
