% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS.MS.Multiple.R
\name{MS.MS.Multiple}
\alias{MS.MS.Multiple}
\title{Protein MS/MS library construction.}
\usage{
MS.MS.Multiple(listSeq, native = FALSE, Metal, Nmetal, Adduct)
}
\arguments{
\item{listSeq}{Output obtain from \code{\link{ProcessMALDIMS}}.}

\item{native}{If TRUE the library contains metal-protein complexes}

\item{Metal}{Metal ion considered (e.g. Zn, Cd)}

\item{Nmetal}{Number of metal ions}

\item{Adduct}{Adduct}
}
\value{
MS/MS ions corresponding to y- and b-serie from permuted Cys-protein sequences.
}
\description{
Calculates y- and b-serie ions for the permuted protein sequences.
}
