% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msms.batch.R
\name{msms.batch}
\alias{msms.batch}
\title{Produce a decoy database}
\usage{
msms.batch(
  sequence,
  missed,
  tol,
  mzML = TRUE,
  label = "Decoy",
  Score,
  plot = FALSE,
  write.csv = FALSE,
  write.csvb = TRUE
)
}
\arguments{
\item{sequence}{protein sequence}

\item{missed}{Control enzymatic digestion}

\item{tol}{mass error tolerance}

\item{mzML}{If TRUE, the file input is in mzML file format}

\item{label}{Name to save}

\item{Score}{Score}

\item{plot}{If TRUE, plot results}

\item{write.csv}{IF TRUE, expor results in csv}

\item{write.csvb}{IF TRUE, expor results in csv}
}
\value{
Decoy database
}
\description{
Produce a decoy database
}
