% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMS.R
\name{MSMS}
\alias{MSMS}
\title{Targeted MS/MS annotation}
\usage{
MSMS(sequence, fragments = "by", IAA = TRUE, N15 = FALSE, custom = list())
}
\arguments{
\item{sequence}{Protein sequence fasta}

\item{fragments}{fragments}

\item{IAA}{IAA}

\item{N15}{N15}

\item{custom}{custom}
}
\value{
Targeted MS/MS annotation
}
\description{
Targeted MS/MS annotation
}
