% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Peak.R
\name{plot_Peak}
\alias{plot_Peak}
\title{Calculates permuted protein sequences according to number of Cys residues present.}
\usage{
plot_Peak(
  file,
  do.plotB = TRUE,
  do.plotA = TRUE,
  axis = c("m/z"),
  norm = TRUE,
  t1 = 1,
  t2 = 10000,
  ID,
  csv = FALSE,
  dir = getwd()
)
}
\arguments{
\item{file}{peak file}

\item{do.plotB}{do plot}

\item{do.plotA}{do plot}

\item{axis}{name axis}

\item{norm}{Normalization intensity}

\item{t1}{mz left}

\item{t2}{mz right}

\item{ID}{Name of file saved}

\item{csv}{save in csv fileformat}

\item{dir}{Directory to save file}
}
\value{
Plot peak
}
\description{
Calculates permuted protein sequences according to number of Cys residues present.
}
