% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeak.R
\name{plotPeak}
\alias{plotPeak}
\title{Calculates permuted protein sequences according to number of Cys residues present.}
\usage{
plotPeak(file, do.plot = TRUE, t1 = 0, t2 = 0, csv = FALSE, orig = TRUE)
}
\arguments{
\item{file}{peak file}

\item{do.plot}{do plot}

\item{t1}{mz left}

\item{t2}{mz right}

\item{csv}{save in csv fileformat}

\item{orig}{Used original file}
}
\value{
Plot peak
}
\description{
Calculates permuted protein sequences according to number of Cys residues present.
}
