% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakBot.R
\name{plotPeakBot}
\alias{plotPeakBot}
\title{Calculates permuted protein sequences according to number of Cys residues present.}
\usage{
plotPeakBot(
  file,
  do.plot = TRUE,
  t1 = 0,
  t2 = 0,
  csv = FALSE,
  orig = FALSE,
  nameTry = nameTry,
  dir = "C:"
)
}
\arguments{
\item{file}{peak file}

\item{do.plot}{do plot}

\item{t1}{mz left}

\item{t2}{mz right}

\item{csv}{save in csv fileformat}

\item{orig}{Used original file}

\item{nameTry}{name to save}

\item{dir}{directory to save}
}
\value{
Plot peak
}
\description{
Calculates permuted protein sequences according to number of Cys residues present.
}
