% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF_bootstrapping.R
\name{PMF_bootstrapping}
\alias{PMF_bootstrapping}
\title{Determine confidence interval by bootstraping}
\usage{
PMF_bootstrapping(
  data,
  R = 999,
  sequence,
  TheoreticalMass,
  Mw,
  ScorePMF,
  missed
)
}
\arguments{
\item{data}{Data}

\item{R}{NUmber of bootstraps}

\item{sequence}{protein sequence}

\item{TheoreticalMass}{Output obtained from \code{\link{Protein.to.Peptide}}.}

\item{Mw}{Molecular weight}

\item{ScorePMF}{Score value}

\item{missed}{Number of missed proteolytic cleavages}
}
\value{
Determine confidence interval by bootstraping for PMF MALDI-MS.
}
\description{
Determine confidence interval by bootstraping
}
