% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prob_score.R
\name{Prob_score}
\alias{Prob_score}
\title{Calculate score probability threshold}
\usage{
Prob_score(maldiTOF, TheoreticalMass, mzw = 2, t2 = 3500, t1 = 500)
}
\arguments{
\item{maldiTOF}{Output obtained from \code{\link{annotationMALDI_Bot}}.}

\item{TheoreticalMass}{Output obtained from \code{\link{Protein.to.Peptide}}.}

\item{mzw}{Mass error window.}

\item{t2}{Left mass}

\item{t1}{Right mass}
}
\value{
Calculate score probability threshold for PMF MALDI-MS.
}
\description{
Calculate score probability threshold
}
