% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessESI.R
\name{ProcessESI}
\alias{ProcessESI}
\title{Pre-process ESI-MS spectra.}
\usage{
ProcessESI(
  file,
  amp.Th = 0.003,
  t1 = 300,
  t2 = 1800,
  norm = TRUE,
  QC1 = FALSE,
  QC2 = FALSE,
  comb = "comb3",
  SNR.Th = 2,
  plotDetection = FALSE,
  plotProcess = FALSE,
  plotOriginal = TRUE,
  xy = FALSE,
  dir = "C:"
)
}
\arguments{
\item{file}{Experimental MS file.}

\item{amp.Th}{Parameter controlling the peak detection.}

\item{t1}{Select lower m/z limit.}

\item{t2}{Select higher m/z limit.}

\item{norm}{Normalize MS spectra. Default to TRUE.}

\item{QC1}{Generate quality control type 1 plot. Default to FALSE.}

\item{QC2}{Generate quality control type 2 plot. Default to FALSE.}

\item{comb}{Select pre-processing method.}

\item{SNR.Th}{Signal-to-noise ratio. Default to 3.}

\item{plotDetection}{Plot detected peaks after peak picking. Default to FALSE.}

\item{plotProcess}{Plot processed MS spectra. Default to FALSE.}

\item{plotOriginal}{Plot raw MS spectra. Default to TRUE.}

\item{xy}{MS spectra in xy file?. Default to FALSE.}

\item{dir}{Directory where to save plots. Default to C:.}
}
\value{
Processed ESI-MS spectra.
}
\description{
Pre-processing and peak picking for ESI-MS spectra.
}
