% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessESI_Decon.R
\name{ProcessESI_Decon}
\alias{ProcessESI_Decon}
\title{Pre-process and deconvolved ESI-MS spectra with simulation algorithm}
\usage{
ProcessESI_Decon(
  file,
  t1 = 300,
  t2 = 1800,
  norm = TRUE,
  comb = "none",
  plotDetection = FALSE,
  plotProcess = FALSE,
  plotOriginal = TRUE,
  xy = FALSE,
  dir = "C:",
  chargemin = 1,
  chargemax = 8,
  maxMass = 8000,
  minMass = 5000,
  FWHM = 2,
  interval = 5,
  Wa = 0.5,
  top = 100,
  SNR.Th = 2
)
}
\arguments{
\item{file}{Experimental MS file.}

\item{t1}{Select lower m/z limit.}

\item{t2}{Select higher m/z limit.}

\item{norm}{Normalize MS spectra. Default to TRUE.}

\item{comb}{Select pre-processing method.}

\item{plotDetection}{Plot detected peaks after peak picking. Default to FALSE.}

\item{plotProcess}{Plot processed MS spectra. Default to FALSE.}

\item{plotOriginal}{Plot raw MS spectra. Default to TRUE.}

\item{xy}{MS spectra in xy file?. Default to FALSE.}

\item{dir}{Directory where to save plots. Default to C:.}

\item{chargemin}{Minimum charge state}

\item{chargemax}{Maximum charge state}

\item{maxMass}{Maximum mass deconvoluted}

\item{minMass}{minimum mass deconvoluted}

\item{FWHM}{FWHM for deconvolution}

\item{interval}{mass interval for deconvolution}

\item{Wa}{mass error for deconvolution}

\item{top}{Top peaks selected for deconvolution}

\item{SNR.Th}{Signal-to-noise ratio. Default to 3.}
}
\value{
Processed and deconvolved ESI-MS spectra.
}
\description{
Pre-process and deconvolved ESI-MS spectra with simulation algorithm
}
