% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessMALDIMS.R
\name{ProcessMALDIMS}
\alias{ProcessMALDIMS}
\title{Pre-process MALDI-MS spectra.}
\usage{
ProcessMALDIMS(
  file,
  t1 = 3000,
  t2 = 8000,
  comb = "comb3",
  SNR.Th = SNR.Th,
  amp.Th = amp.Th,
  xy = FALSE,
  plotProcess = FALSE,
  norm = TRUE,
  plotDetection = FALSE,
  plotOriginal = TRUE,
  dir = getwd()
)
}
\arguments{
\item{file}{Experimental MALDI-MS file.}

\item{t1}{Select lower m/z limit.}

\item{t2}{Select higher m/z limit.}

\item{comb}{Select pre-processing method.}

\item{SNR.Th}{Signal-to-noise ratio. Default to 3.}

\item{amp.Th}{Parameter controlling the peak detection.}

\item{xy}{MS spectra in xy file?. Default to FALSE.}

\item{plotProcess}{Plot processed MS spectra. Default to FALSE.}

\item{norm}{Normalize MS spectra. Default to TRUE.}

\item{plotDetection}{Plot detected peaks after peak picking. Default to FALSE.}

\item{plotOriginal}{Plot raw MS spectra. Default to TRUE.}

\item{dir}{Directory where to save plots. Default to C:.}
}
\value{
Processed MALDI-MS spectra.
}
\description{
Pre-processing and peak picking for ESI-MS spectra.
}
