% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein.to.List.R
\name{Protein.to.List}
\alias{Protein.to.List}
\title{Library generation of chemical protein Cys-labelled residues.}
\usage{
Protein.to.List(
  sequence,
  IAA = FALSE,
  IAM = FALSE,
  ET = FALSE,
  NEM = FALSE,
  LABEL = TRUE,
  monoisotopic = FALSE
)
}
\arguments{
\item{sequence}{Protein sequence.}

\item{IAA}{Selection of chemical labelling reagent used: Iodoacetic acid.}

\item{IAM}{Selection of chemical labelling reagent used: Iodoacetamide.}

\item{ET}{Selection of chemical labelling reagent used: Ethyl iodoacetate.}

\item{NEM}{Selection of chemical labelling reagent used: NEM.}

\item{LABEL}{Chemical labelling. Default to TRUE.}

\item{monoisotopic}{Monoisotopic mass considered for library construction. Default to FALSE.}
}
\value{
MS library generated.
}
\description{
Generates theoretical list of Cys-labelled residues with a particular chemical labelling reagent.
}
