% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein.To.Peptide.R
\name{Protein.to.Peptide}
\alias{Protein.to.Peptide}
\title{Proteolytic digestion of a partial chemical labelled protein.}
\usage{
Protein.to.Peptide(sequence, missed = 3, enzym = "trypsin")
}
\arguments{
\item{sequence}{Protein sequence.}

\item{missed}{Missed cleavage sites produced by proteolytic enzyme. Default to 2.}

\item{enzym}{Proteolytic enzyme used in the experiment. Default to Trypsin.}
}
\value{
List of peptides with incomplete chemically labelled Cys-protein residues.
}
\description{
Generates a MS library containing peptides harbouring incomplete chemical Cys-labelled residues.
}
\examples{
\dontrun{
MT2a<-("MDPNCSCAAGDSCTCAGSCKCKECKCTSCKKSCCSCCPVGCAKCAQGCICKGASDKCSCCA")
TheoreticalMass<-Protein.to.Peptide(MT2a)

}
}
