% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_deconvolution.R
\name{Score_deconvolution}
\alias{Score_deconvolution}
\title{Scoring function for peptide mass fingerprint MALDI-MS experiment.}
\usage{
Score_deconvolution(
  ESIprocessed,
  intervaLo = 3000,
  intervaLo2 = 2000,
  loess = FALSE,
  span = 0.002,
  mz1 = NULL,
  mz2 = NULL
)
}
\arguments{
\item{ESIprocessed}{Processed ESI-MS spectra.}

\item{intervaLo}{mass interval for deconvolution}

\item{intervaLo2}{mass interval for deconvolution}

\item{loess}{If TRUE, loess algorithm is used.}

\item{span}{span controlling the loess}

\item{mz1}{Left mass}

\item{mz2}{Right mass}
}
\value{
Annotation score for PMF MALDI-MS.
}
\description{
Scoring function for peptide mass fingerprint MALDI-MS experiment.
}
