% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreMSMS.R
\name{ScoreMSMS}
\alias{ScoreMSMS}
\title{Scoring function for MS/MS spectra.}
\usage{
ScoreMSMS(
  theoryMultiple,
  final,
  tol,
  exp,
  Score = 4,
  write.csv = TRUE,
  write.csvb = TRUE
)
}
\arguments{
\item{theoryMultiple}{Theoretical MS/MS library.}

\item{final}{Output Theoretical MS/MS library.}

\item{tol}{Mass error}

\item{exp}{Experimental MS/MS file.}

\item{Score}{Select the type of score using for finding best prediction following the next order: These are number of ions matched, percentage peaks matched and percentage of ions matched.}

\item{write.csv}{Write output.}

\item{write.csvb}{Write output.}
}
\value{
Annotation score for MS/MS spectra
}
\description{
Scoring function for MS/MS spectra.
}
