% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoresPMF.R
\name{ScoresPMF}
\alias{ScoresPMF}
\title{Scoring function for peptide mass fingerprint MALDI-MS experiment.}
\usage{
ScoresPMF(
  sequence,
  maldiTOF,
  TheoreticalMass,
  Total,
  Mw,
  write = TRUE,
  save.name,
  missed,
  enzym = "trypsin"
)
}
\arguments{
\item{sequence}{Protein sequence.}

\item{maldiTOF}{Output obtained from \code{\link{annotationMALDI_Bot}}.}

\item{TheoreticalMass}{Output obtained from \code{\link{Protein.to.Peptide}}.}

\item{Total}{Number of amino acids in the protein sequence.}

\item{Mw}{Molecular weight of the protein assayed.}

\item{write}{Write output with PMF score.}

\item{save.name}{Name to save the output results.}

\item{missed}{Number of missed proteolytic cleavages}

\item{enzym}{Enzyme used for proteolysis.}
}
\value{
Annotation score for PMF MALDI-MS.
}
\description{
Scoring function for peptide mass fingerprint MALDI-MS experiment.
}
