% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceCounting.R
\name{SequenceCounting}
\alias{SequenceCounting}
\title{Sequence coveraged from Bottom-up MALDI-MS experiment.}
\usage{
SequenceCounting(seq, maldiTOF, missed, enzym = "trypsin")
}
\arguments{
\item{seq}{Protein sequence.}

\item{maldiTOF}{Output obtained from \code{\link{annotationMALDI_Bot}}.}

\item{missed}{Missed cleavage sites produced by proteolytic enzyme. Default to 2.}

\item{enzym}{Proteolytic enzyme used in the experiment. Default to Trypsin.}
}
\value{
Protein sequence coverage for bottom-up MALDI-MS experiment.
}
\description{
Calculates the protein sequence coverage provided by bottom-up MALDI-MS experiments
}
\examples{
\dontrun{
MT2a<-("MDPNCSCAAGDSCTCAGSCKCKECKCTSCKKSCCSCCPVGCAKCAQGCICKGASDKCSCCA")
TheoreticalMass<-Protein.to.Peptide(MT2a)
ProcessMALDIMS<-ProcessMALDIMS(file,t1=500,t2=3500,plotProcess=TRUE,plotDetection=TRUE,plotOriginal = TRUE,comb = "comb3")
maldiTOF<-annotationMALDI_Bot(ProcessMALDIMS,TheoreticalMass,mzw = 1,plot=TRUE,plotIndi=TRUE,orig=TRUE,write=TRUE,In=10)
maldiTOF.Amp(MT2a,maldiTOF)
CountAAPMFF<-SequenceCounting(MT2a,maldiTOF)
}
}
