% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.MSMS.Multiple_B.R
\name{target.MSMS.Multiple_B}
\alias{target.MSMS.Multiple_B}
\title{Annotate top-down MS/MS experiments from multiple Cys-protein.}
\usage{
target.MSMS.Multiple_B(
  theoryMultiple,
  exp,
  t = 0.4,
  b = 5,
  write = FALSE,
  plot = FALSE,
  supress = FALSE,
  xlim = NULL
)
}
\arguments{
\item{theoryMultiple}{Output obtain from \code{\link{MS.MS.Multiple}}.}

\item{exp}{Experimental MS/MS file.}

\item{t}{Mass-to-charge window (Da) for the annotation. Default to 0.4.}

\item{b}{Threshold for relative intensity. Default to 5.}

\item{write}{Output the results}

\item{plot}{plot the MS/MS spectra}

\item{supress}{Include MS/MS ions in the output plot. Default to FALSE.}

\item{xlim}{Mass range considered for the annotation.}
}
\value{
Peaks annotated in the top-down MS/MS spectra.
}
\description{
Annotate top-down MS/MS experiments from multiple Cys-protein.
}
