% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetESI.R
\name{TargetESI}
\alias{TargetESI}
\title{Annotate ESI-MS spectra}
\usage{
TargetESI(parent, Theoretical.esi, Adduct, mzw = 1)
}
\arguments{
\item{parent}{Parent mass}

\item{Theoretical.esi}{Theoretical library}

\item{Adduct}{Adduct}

\item{mzw}{Mass-to-charge window (Da) for the annotation. Default to 1.}
}
\value{
Peaks annotated in the ESI-MS spectra.
}
\description{
Annotate ESI-MS spectra
}
