% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetMSMS_bottom.R
\name{targetMSMS_bottom}
\alias{targetMSMS_bottom}
\title{Annotate ESI-MS/MS spectra}
\usage{
targetMSMS_bottom(
  Deconv_CysMpro,
  theory,
  mzw = 1,
  b = 5,
  plot = FALSE,
  supress = FALSE,
  xlim = c(100, 2500),
  label = ""
)
}
\arguments{
\item{Deconv_CysMpro}{Deconvolved ESI-MS spectra.}

\item{theory}{Theoretical ESI-MS/MS library}

\item{mzw}{Mass-to-charge window (Da) for the annotation. Default to 1.}

\item{b}{Intensity threshold}

\item{plot}{If TRUE, plot the results}

\item{supress}{Control the plot options}

\item{xlim}{Select m/z range}

\item{label}{Label the results}
}
\value{
Peaks annotated in the MALDI-MS spectra.
}
\description{
Annotate ESI-MS/MS spectra
}
