% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topdown.R
\name{Topdown}
\alias{Topdown}
\title{Annotate a top-down MS/MS spectra}
\usage{
Topdown(
  experimental,
  labelling = FALSE,
  mz1 = 500,
  mz2 = 3000,
  max.charge,
  tol,
  sequence,
  t,
  b,
  SNR.Th,
  peak = TRUE,
  plot = FALSE,
  maxMass,
  parent,
  mzw,
  Metal = c("Zn"),
  NbindingMetal = 7,
  Maxcharge = 6,
  Adduct = NULL
)
}
\arguments{
\item{experimental}{Experimental MS/MS spectra}

\item{labelling}{If TRUE, a CYs labelling reagent is used.}

\item{mz1}{left mass}

\item{mz2}{right mass}

\item{max.charge}{maximum charge state for deconvolution}

\item{tol}{mass tolerance}

\item{sequence}{Protein sequence}

\item{t}{Mass error window.}

\item{b}{Filter by intensity}

\item{SNR.Th}{SNR threshold}

\item{peak}{if TRUE peak picking is performed}

\item{plot}{if TRUE, plot is produced}

\item{maxMass}{maximum mass for deconvolution}

\item{parent}{parent ion selected for MS/MS}

\item{mzw}{mass error window to annotate the parent ion}

\item{Metal}{Metal}

\item{NbindingMetal}{Number of metal ions}

\item{Maxcharge}{maximum charge state for theoretical database}

\item{Adduct}{If NULL, proton is the adduct considered}
}
\value{
Annotate top-down MS/MS spectra
}
\description{
Annotate a top-down MS/MS spectra
}
