% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topdown_batch.R
\name{Topdown_batch}
\alias{Topdown_batch}
\title{Annotate a set of top-down MS/MS spectra}
\usage{
Topdown_batch(
  mzML = TRUE,
  xy = FALSE,
  mz1 = 500,
  mz2 = 3000,
  tol = 0.15,
  txt = FALSE,
  Metal = c("Zn"),
  SNR.Th = 2,
  peak = FALSE,
  plot = FALSE,
  dir = getwd(),
  sequence,
  b,
  t,
  maxMass,
  mzw,
  NbindingMetal,
  Maxcharge,
  label = "CID"
)
}
\arguments{
\item{mzML}{if TRUE, the experimental file is in mzML file format}

\item{xy}{if TRUE, the experimental file is in xy file format}

\item{mz1}{left mass}

\item{mz2}{right mass}

\item{tol}{mass tolerance}

\item{txt}{if TRUE, the experimental file is in txt file format}

\item{Metal}{Metal}

\item{SNR.Th}{SNR threshold}

\item{peak}{if TRUE peak picking is performed}

\item{plot}{if TRUE, plot is produced}

\item{dir}{Where to save results}

\item{sequence}{Protein sequence}

\item{b}{Filter by intensity}

\item{t}{Mass error window.}

\item{maxMass}{maximum mass for deconvolution}

\item{mzw}{mass error window to annotate the parent ion}

\item{NbindingMetal}{Number of metal ions}

\item{Maxcharge}{maximum charge state for deconvolution}

\item{label}{Label the results}
}
\value{
Annotate a set of top-down MS/MS spectra
}
\description{
Annotate a set of top-down MS/MS spectra
}
