% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate_score.R
\name{Validate_score}
\alias{Validate_score}
\title{Validate the scoring function for peptide mass fingerprint MALDI-MS experiment.}
\usage{
Validate_score(
  ProcessMALDIMS,
  TheoreticalMass,
  mzw,
  nAA,
  threshold,
  Mw,
  sequence,
  Npermu = 10000,
  In,
  prob = FALSE,
  score_prob_threshold,
  missed
)
}
\arguments{
\item{ProcessMALDIMS}{Processed MALDI-MS spectrum Protein sequence.}

\item{TheoreticalMass}{Output obtained from \code{\link{Protein.to.Peptide}}.}

\item{mzw}{Mass error window.}

\item{nAA}{Number of amino acids in the protein sequence.}

\item{threshold}{threshold for scoring}

\item{Mw}{Molecular weight of the protein assayed.}

\item{sequence}{Protein sequence}

\item{Npermu}{Number of permutations}

\item{In}{In for annotationMALDI_Bot function}

\item{prob}{If TRUE returns the score calculated with binomial function}

\item{score_prob_threshold}{Threshold for probability}

\item{missed}{Number of missed proteolytic cleavages}
}
\value{
Annotation score for PMF MALDI-MS.
}
\description{
Validate the scoring function for peptide mass fingerprint MALDI-MS experiment.
}
