% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_topdown.R
\name{Validation_topdown}
\alias{Validation_topdown}
\title{Validation top-down MS/MS results}
\usage{
Validation_topdown(
  ev50_true,
  sequence,
  mzw = 2,
  Deconv_CysMpro,
  intervalmz = 3000,
  Npermu = 1000,
  Nmetal = 7,
  Metal = c("Zn")
)
}
\arguments{
\item{ev50_true}{results from topdown.R function}

\item{sequence}{Protein sequence}

\item{mzw}{Mass error window.}

\item{Deconv_CysMpro}{Deconvolved ESI-MS}

\item{intervalmz}{interval m/z}

\item{Npermu}{Number of permutations}

\item{Nmetal}{Number of metal ions}

\item{Metal}{Metal}
}
\value{
Validation top-down MS/MS results
}
\description{
Validation top-down MS/MS results
}
